Issue #SB-597 feat: Review rework
